requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
psycopg2-binary==2.9.3
Pillow==9.0.1